+++ CHANGELOG.rst
@@ -339,3 +339,13 @@ Changelog
 * Fixed the issue in 1.16.0 where plugin names would log as "None".
 * Added sample task CSV file showing the required layout for ``sb-csv``
 * Updated documentation for pet care and tasks.
+
+1.17.2 (Tuesday January 17, 2017) 
+-----------------------------------------
+
+* Very minor changes to PyPI metadata and README file.
+* Made the update and creation of the notification panel optional. Use the
+  `use-notification-panel` command line argument to control the behaviour.
+* Made the task tags configurable. Use the ``--tags`` argument to pass a comma
+  separated list of tags that will be used with any task created in Habitica. If
+  you use ``--tags ""`` then no tags will be created or applied.